==> data process/1_ais_dataset_process_hs.ipynb <==
## 1 AIS dataset process 

# import libraries

import pandas as pd
import numpy as np

import matplotlib
import matplotlib.pyplot as plt
from matplotlib import figure
from matplotlib.colors import ListedColormap
from matplotlib import cm

from PIL import Image
import requests
import math
from scipy.interpolate import interp1d
import datetime
from datetime import datetime
import matplotlib.collections as colors


to_process_data_path = r'C:\Users\Public\Documents\AIS Models\dataset\01_split_2020_v0_ais_dataset\data_2020_01.csv'
to_store_data_path = r'C:\Users\Public\Documents\AIS Models\dataset\02_clean_2020_v1_ais_process\data_2020_01_v1_out.csv'

# columns = ['mmsi', 'unixtime', 'latitude', 'longitude', 'heading', 'sog', 'nav_status', 'cog', 'rot']


# Process the input AIS data_from split sets ; Renaming columns

def process_data(file_path): 

    ais_raw_data = pd.read_csv(file_path)

    ## use if delimiters present in the data file 
    # ais_raw_data = pd.read_csv(file_path , delimiter=';') 
    
    # # converting date_time column to standard date time format 
    # # produce a date_time_stamp column
    # ais_raw_data['date_time_stamp'] = pd.to_datetime(ais_raw_data[next(col for col in ais_raw_data.columns if 'time' in col)], unit= 's')

    s1_ais_raw_data = ais_raw_data.copy().drop_duplicates()
    print( 'The shape of AIS Dataset (after dropping duplicate rows is: ', s1_ais_raw_data.shape)
    print('Number of rows dropped:', (ais_raw_data.shape[0] - s1_ais_raw_data.shape[0]))

    # Changing the name of certain (key) columns:
    # Key identified columns: mmsi, lat, lon, head, sog, cog, nav

    s1_ais_raw_data.rename (columns= { next(col for col in ais_raw_data.columns if 'mmsi' in col): 'mmsi',
    next(col for col in ais_raw_data.columns if 'lat' in col): 'lat',
    next(col for col in ais_raw_data.columns if 'lon' in col): 'lon',
    next(col for col in ais_raw_data.columns if 'head' in col): 'heading',
    next(col for col in ais_raw_data.columns if 'sog' in col): 'sog',
    next(col for col in ais_raw_data.columns if 'cog' in col): 'cog',
    next(col for col in ais_raw_data.columns if 'nav' in col): 'nav_status'
    },inplace= True)
    return s1_ais_raw_data



# Support the lat lon rounding up and down for Max and Min limits respectively  (used with visualizing / plotting data)
def round_up(float_num):
    rounded_up = math.ceil(float_num * 1000) / 1000
    return rounded_up
def round_down(float_num):
    rounded_down = math.floor(float_num * 1000) / 1000
    return rounded_down



# Get lat long limits for BoundBox required limits approximation (used with visualizing / plotting data)
def get_lon_lat_limits (df):
    longitude_column = next(col for col in df.columns if 'lon' in col)
    latitude_column = next(col for col in df.columns if 'lat' in col)

    df[longitude_column] = df[longitude_column].str.replace(',', '.').astype(float)
    df[latitude_column]= df[latitude_column].str.replace(',', '.').astype(float)

    lat_max  = df[latitude_column].max()
    lat_min  = df[latitude_column].dropna().min()
    
    lon_max = df[longitude_column].max()
    lon_min  = df[longitude_column].dropna().min()
    
    # print('MH max longitude:', lon_max)
    # print('MH min longitude:', lon_min)
    # print('MH max latitude:', lat_max)
    # print('MH min latitude:', lat_min)

    #BBox limits (excluding margin) 
    b_lon_max = round_up (lon_max)
    b_lon_min = round_down(lon_min)
    b_lat_max = round_up (lat_max)
    b_lat_min = round_down(lat_min)

    print('MH max longitude:', b_lon_max)
    print('MH min longitude:', b_lon_min)
    print('MH max latitude:', b_lat_max)
    print('MH min latitude:', b_lat_min)
 
    return {'lon_max':b_lon_max,'lon_min': b_lon_min,'lat_max': b_lat_max, 'lat_min':b_lat_min}    # Dictionary of lon lat limits


# Processing data

# 2019_oslo_ais_data
ais_data_v1 = process_data(to_process_data_path)

# map_limits = get_lon_lat_limits(ais_data_v1) # Determining map limits for BBox 
# print(map_limits)

# DATA PROCESSING - STEP 2

## FUNCTION DEFINITION

# GENERATING UNIQUE VOYAGES (MMSI --> TIME STAMP DIFFENCE AND )
# And assignment of MMSI_VOYAGE NUMBER STAMPS for further processing--> Step 3: filtering valid voyages according to minimum limit of datapoint counts.  

def process_voyage_generation (df):
    #list of unique mmsi_tags
    mmsi_tags = df['mmsi'].unique()
    df['date_time_stamp'] = pd.to_datetime(df['date_time_stamp'])
    ais_data_clean = pd.DataFrame() # Processing step 2 _time as assigning Voyage number based on mmsi and time start stamps
    count = 0
    for mmsi_tag in mmsi_tags:
        count += 1
        print(count)
        step_data_time_raw = df[df['mmsi'] == mmsi_tag] # Selecting a mmsi
        step_data_time_raw = step_data_time_raw.sort_values('date_time_stamp') # Time Sorting all points of the  mmsi 
        step_data_time =  step_data_time_raw.reset_index(drop=True) # Reseting index after time sort for the mmsi

        # use if needed with dataset conversions and cog missing value handling
        # convert to 'date time object' from 'object'
        # step_data_time['date_time_stamp'] = pd.to_datetime(step_data_time['date_time_stamp'])
        # Convering cog from 'object' to 'numeric' values
        # cog_stamps = pd.to_numeric(step_data_time['cog'], errors='coerce')

        time_stamps = step_data_time['date_time_stamp'] # Creating a list of time_stamps 
        # THRESHOLD time for new voyage identification : 10 minutes = 600 seconds _based on inference from the AIS message time difference. 
        threshold_time = 600 # Seconds (10 minutes)
        voyage_id_stamp = int(str(mmsi_tag) + str(step_data_time['unixtime(timestamp)'][0])) # creating a unique mmsi _ time stamp for voyage
        step_data_time.loc[0,'mmsi_voyage_num'] = voyage_id_stamp 
        
        # loop through the time stamps and check if the difference is greater than the threshold
        for i in range(1, len(time_stamps)):
            time_diff = int((datetime.strptime(time_stamps[i].to_pydatetime().strftime("%Y-%m-%d %H:%M:%S+00:00"), "%Y-%m-%d %H:%M:%S+00:00") - datetime.strptime(time_stamps[i-1].to_pydatetime().strftime("%Y-%m-%d %H:%M:%S+00:00"), "%Y-%m-%d %H:%M:%S+00:00")).total_seconds()) # or abs(cog_stamps[i] - cog_stamps[i-10]) < 60 : #or math.isnan()))  :
            if time_diff > threshold_time:
                voyage_id_stamp = int(str(mmsi_tag) + str(step_data_time['unixtime(timestamp)'][i]))
                step_data_time.loc[i,'mmsi_voyage_num'] = voyage_id_stamp 

            else:
                step_data_time.loc[i, 'mmsi_voyage_num'] = voyage_id_stamp
        
        ais_data_clean = pd.concat([ais_data_clean, step_data_time], ignore_index = True)
    return ais_data_clean
            

# DATA PROCESSING - STEP 2
ais_data_v2 = process_voyage_generation(ais_data_v1)

# Read clean ais dataset
ais_data_oslo_600 = ais_data_v2.copy()
# DATA PROCESSING- COG Values 
ais_data_oslo_600_with_cog = ais_data_oslo_600.copy()


# ## For use in case COG value are a mix of dot and comma-separated: 

# # Replace commas with dots
# ais_data_oslo_600['cog'] = ais_data_oslo_600['cog'].str.replace(',', '.')
# # Convert to float
# ais_data_oslo_600['cog'] = pd.to_numeric(ais_data_oslo_600['cog'], errors='coerce')
# # Replace NaN with original values
# ais_data_oslo_600['cog'] = ais_data_oslo_600['cog'].fillna(ais_data_oslo_600_with_cog['cog'])

ais_data_oslo_600.to_csv(to_store_data_path, index= False)
