==> data process/6_ais_dataset_process_hs.ipynb <==
## 6 Training dataset creation - concatenating months (Jan to Oct) 

import pandas as pd
import numpy as np

# # Save the datasets as NumPy arrays : with 6 key features standardized ('lat', 'lon', 'sog', 'cog_sin', 'cog_cos', 'NDC_nm', in_TSS, Vessel_TSS  )
to_store_data_path_X_train = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_8_output\1_10_train_rnn\data_2019_train_X_v8_out.np'
to_store_data_path_Y_lat_train = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_8_output\1_10_train_rnn\data_2019_train_Y_lat_v8_out.np'
to_store_data_path_Y_lon_train = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_8_output\1_10_train_rnn\data_2019_train_Y_lon_v8_out.np'
to_store_data_path_Y_train = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_8_output\1_10_train_rnn\data_2019_train_Y_v8_out.np'

read_1_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_01_v5_out\data_2019_01_X.npy'
read_2_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_02_v5_out\data_2019_02_X.npy'
read_3_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_03_v5_out\data_2019_03_X.npy'
read_4_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_04_v5_out\data_2019_04_X.npy'
read_5_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_05_v5_out\data_2019_05_X.npy'
read_6_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_06_v5_out\data_2019_06_X.npy'
read_7_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_07_v5_out\data_2019_07_X.npy'
read_8_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_08_v5_out\data_2019_08_X.npy'
read_9_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_09_v5_out\data_2019_09_X.npy'
read_10_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_10_v5_out\data_2019_10_X.npy'

# read_11_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_11_v5_out\data_2019_11_X.npy'

# read_12_X = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_12_v5_out\data_2019_12_X.npy'


read_1_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_01_v5_out\data_2019_01_Y_lat.npy'
read_2_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_02_v5_out\data_2019_02_Y_lat.npy'
read_3_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_03_v5_out\data_2019_03_Y_lat.npy'
read_4_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_04_v5_out\data_2019_04_Y_lat.npy'
read_5_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_05_v5_out\data_2019_05_Y_lat.npy'
read_6_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_06_v5_out\data_2019_06_Y_lat.npy'
read_7_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_07_v5_out\data_2019_07_Y_lat.npy'
read_8_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_08_v5_out\data_2019_08_Y_lat.npy'
read_9_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_09_v5_out\data_2019_09_Y_lat.npy'
read_10_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_10_v5_out\data_2019_10_Y_lat.npy'

# read_11_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_11_v5_out\data_2019_11_Y_lat.npy'

# read_12_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_12_v5_out\data_2019_12_Y_lat.npy'


read_1_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_01_v5_out\data_2019_01_Y_lon.npy'
read_2_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_02_v5_out\data_2019_02_Y_lon.npy'
read_3_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_03_v5_out\data_2019_03_Y_lon.npy'
read_4_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_04_v5_out\data_2019_04_Y_lon.npy'
read_5_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_05_v5_out\data_2019_05_Y_lon.npy'
read_6_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_06_v5_out\data_2019_06_Y_lon.npy'
read_7_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_07_v5_out\data_2019_07_Y_lon.npy'
read_8_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_08_v5_out\data_2019_08_Y_lon.npy'
read_9_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_09_v5_out\data_2019_09_Y_lon.npy'
read_10_Y_lon = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_10_v5_out\data_2019_10_Y_lon.npy'

# read_11_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_11_v5_out\data_2019_11_Y_lat.npy'

# read_12_Y_lat = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_12_v5_out\data_2019_12_Y_lat.npy'


read_1_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_01_v5_out\data_2019_01_Y.npy'
read_2_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_02_v5_out\data_2019_02_Y.npy'
read_3_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_03_v5_out\data_2019_03_Y.npy'
read_4_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_04_v5_out\data_2019_04_Y.npy'
read_5_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_05_v5_out\data_2019_05_Y.npy'
read_6_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_06_v5_out\data_2019_06_Y.npy'
read_7_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_07_v5_out\data_2019_07_Y.npy'
read_8_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_08_v5_out\data_2019_08_Y.npy'
read_9_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_09_v5_out\data_2019_09_Y.npy'
read_10_Y = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_5_output\data_2019_10_v5_out\data_2019_10_Y.npy'

# Load the datasets from the saved files

data_1_X = np.load(read_1_X)
data_2_X = np.load(read_2_X)
data_3_X = np.load(read_3_X)
data_4_X = np.load(read_4_X)
data_5_X = np.load(read_5_X)
data_6_X = np.load(read_6_X)
data_7_X = np.load(read_7_X)
data_8_X = np.load(read_8_X)
data_9_X = np.load(read_9_X)
data_10_X = np.load(read_10_X)

# data_11_X = np.load(read_11_X)
# data_12_X = np.load(read_12_X)

# Load the datasets from the saved files

data_1_Y_lat = np.load(read_1_Y_lat)
data_2_Y_lat = np.load(read_2_Y_lat)
data_3_Y_lat = np.load(read_3_Y_lat)
data_4_Y_lat = np.load(read_4_Y_lat)
data_5_Y_lat = np.load(read_5_Y_lat)
data_6_Y_lat = np.load(read_6_Y_lat)
data_7_Y_lat = np.load(read_7_Y_lat)
data_8_Y_lat = np.load(read_8_Y_lat)
data_9_Y_lat = np.load(read_9_Y_lat)
data_10_Y_lat = np.load(read_10_Y_lat)

# data_11_Y_lat = np.load(read_11_Y_lat)
# data_12_Y_lat = np.load(read_12_Y_lat)

# Load the datasets from the saved files

data_1_Y_lon = np.load(read_1_Y_lon)
data_2_Y_lon = np.load(read_2_Y_lon)
data_3_Y_lon = np.load(read_3_Y_lon)
data_4_Y_lon = np.load(read_4_Y_lon)
data_5_Y_lon = np.load(read_5_Y_lon)
data_6_Y_lon = np.load(read_6_Y_lon)
data_7_Y_lon = np.load(read_7_Y_lon)
data_8_Y_lon = np.load(read_8_Y_lon)
data_9_Y_lon = np.load(read_9_Y_lon)
data_10_Y_lon = np.load(read_10_Y_lon)

# data_11_Y_lon = np.load(read_11_Y_lat)
# data_12_Y_lon = np.load(read_12_Y_lat)

# Load the datasets from the saved files

data_1_Y = np.load(read_1_Y)
data_2_Y = np.load(read_2_Y)
data_3_Y = np.load(read_3_Y)
data_4_Y = np.load(read_4_Y)
data_5_Y = np.load(read_5_Y)
data_6_Y = np.load(read_6_Y)
data_7_Y = np.load(read_7_Y)
data_8_Y = np.load(read_8_Y)
data_9_Y = np.load(read_9_Y)
data_10_Y = np.load(read_10_Y)


X_train_frames = (data_1_X, data_2_X, data_3_X, data_4_X, data_5_X, data_6_X, data_7_X, data_8_X, data_9_X, data_10_X)
Y_lat_train_frames = (data_1_Y_lat, data_2_Y_lat, data_3_Y_lat, data_4_Y_lat, data_5_Y_lat, data_6_Y_lat, data_7_Y_lat, data_8_Y_lat, data_9_Y_lat, data_10_Y_lat)
Y_lon_train_frames = (data_1_Y_lon, data_2_Y_lon, data_3_Y_lon, data_4_Y_lon, data_5_Y_lon, data_6_Y_lon, data_7_Y_lon, data_8_Y_lon, data_9_Y_lon, data_10_Y_lon)

Y_train_frames = (data_1_Y, data_2_Y, data_3_Y, data_4_Y, data_5_Y, data_6_Y, data_7_Y, data_8_Y, data_9_Y, data_10_Y)

data_X_train = np.concatenate(X_train_frames, axis=0)
data_Y_lat_train = np.concatenate(Y_lat_train_frames, axis=0)
data_Y_lon_train = np.concatenate(Y_lon_train_frames, axis=0)

data_Y_train = np.concatenate(Y_train_frames, axis=0)


# Save the datasets as NumPy arrays : with 6 key features standardized ('lat', 'lon', 'sog', 'cog_sin', 'cog_cos', 'NDC_nm', in_TSS, Vessel_TSS  )
np.save(to_store_data_path_X_train, data_X_train)
np.save(to_store_data_path_Y_lat_train, data_Y_lat_train)
np.save(to_store_data_path_Y_lon_train, data_Y_lon_train)

np.save(to_store_data_path_Y_train, data_Y_train)
