==> data process/7_ais_dataset_process_encounter_unifrom_ts.ipynb <==
## 7 AIS dataset process - Encounter generator (Not considered towards thesis scope)

# from sklearn.neighbors import BallTree
import numpy as np
import pandas as pd
from haversine import haversine, Unit

pd.options.mode.chained_assignment = None  # default='warn'

to_read_own_vessel_path = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_2_output_unifom_ts\data_2019_02_v2_uniform_ts_out.csv'
to_read_other_vessel_path = r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_6_output_uniform_ts\data_2019_02_v6_out_uniform_ts.csv'
to_store_path =  r'W:\Neeraj_thesis\version2_thesis\dataset\2019\generator_version_10_output_encounter_unifom_ts\data_2019_02_v10_out_encounters_uniform_ts.csv'

ais_own_data = pd.read_csv(to_read_own_vessel_path)
ais_other_data = pd.read_csv(to_read_other_vessel_path)

ais_encounter_data = ais_own_data.copy()

for i in range(1, 11):
    ais_encounter_data[f'{i}_encounter_vessel_mmsi'] = None
    ais_encounter_data[f'{i}_encounter_vessel_lat'] = None
    ais_encounter_data[f'{i}_encounter_vessel_lon'] = None
    
    ais_encounter_data[f'{i}_encounter_distance'] = None
    ais_encounter_data[f'{i}_encounter_sog'] = None
    ais_encounter_data[f'{i}_encounter_cog'] = None
    # ais_encounter_data[f'{i}_encounter_relative_bearing'] = None
    ais_encounter_data[f'{i}_encounter_time'] = None
    

print(ais_own_data.columns)
print(ais_other_data.columns)
print(ais_encounter_data.columns)

# Compute encounters for each vessel

def find_encounters(ais_own_data, ais_other_data, ais_encounter_df, encounter_distance_limit):
    encounter_df = pd.DataFrame()
    vessels = ais_own_data["mmsi"].unique()
#     print(vessels)
    count = 0
    ais_encounter_data = ais_encounter_df.copy()

    
    for vessel in vessels:
        # if count < 3:

        count+=1
        print(count)
        
        vessel_data = ais_own_data[ais_own_data["mmsi"] == vessel].sort_values("unixtime")
        other_vessels_data = ais_other_data[ais_other_data["mmsi"] != vessel]

        for index, row in vessel_data.iterrows():
            enc_count = 1
            # print(index)
            lat, lon = row[["lat", "lon"]]
            other_vessels_data_v1= pd.DataFrame()
            other_vessels_data_v1 = other_vessels_data[other_vessels_data['unixtime'] == row['unixtime']] 
            
            if not other_vessels_data_v1.empty:
                # print(other_vessels_data_v1)
                # print('go on')
                other_vessels_data_v1.loc[:,'distance'] = other_vessels_data_v1.apply(lambda x: haversine((lat, lon), (x['lat'], x['lon']), unit=Unit.KILOMETERS), axis=1)
                # other_vessels_data_v1['distance'] = haversine((lat, lon), (other_vessels_data_v1['lat'], other_vessels_data_v1['lon']), unit= Unit.KILOMETERS)

                other_vessels_data_v2 = other_vessels_data_v1[other_vessels_data_v1['distance'] < encounter_distance_limit]
                
                other_vessels_data_v2 = other_vessels_data_v2.sort_values('distance', ascending=True)


                if not other_vessels_data_v2.empty:
                    for enc_vessel, enc_vessel_data in other_vessels_data_v2.groupby("mmsi"):
                        while enc_count < 8:
                            # print(enc_vessel_data)
                            ais_encounter_data.at[index,f'{enc_count}_encounter_vessel_mmsi'] = enc_vessel_data['mmsi']
                            ais_encounter_data.at[index,f'{enc_count}_encounter_vessel_lat'] = enc_vessel_data['lat']
                            ais_encounter_data.at[index,f'{enc_count}_encounter_vessel_lon'] = enc_vessel_data['lon']
                            ais_encounter_data.at[index,f'{enc_count}_encounter_distance'] = enc_vessel_data['distance']
                            # print(enc_vessel_data['distance'])
                            ais_encounter_data.at[index,f'{enc_count}_encounter_sog'] = enc_vessel_data['sog']
                            ais_encounter_data.at[index,f'{enc_count}_encounter_cog'] = enc_vessel_data['cog']
                            # ais_encounter_data.loc[index,f'{enc_count}_encounter_relative_bearing'] = enc_vessel_data['sog']
                            ais_encounter_data.at[index,f'{enc_count}_encounter_time'] = enc_vessel_data['mmsi']
                            
                            enc_count += 1

                else:
                    continue
                        
            else:
                continue
                        

            
        # else:
        #     break

    return (ais_encounter_data)

        


encounter_distance_limit = 3  # 2 km, for example
# encounter_time_window = 30  # 15 minutes, for example *NOT NEEDED TO BE USED ANY MORE AS CAN EQUATE TIME (ALL FOLLOW A UNIFROM TIME DISTRIBUTION WITH OVERLAPPIG TIME PERIOD)

encounters = find_encounters(ais_own_data = ais_own_data, ais_other_data= ais_other_data, ais_encounter_df = ais_encounter_data,  encounter_distance_limit =encounter_distance_limit)

encounters.to_csv(to_store_path)
